requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
boltons==23.0.0
brotlipy==0.7.0
cachetools==5.3.0
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625807838443/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
charset-normalizer==3.1.0
clldutils==3.19.0
cmake==3.26.0
colorama==0.4.6
colorlog==6.7.0
comm==0.1.2
conda==4.10.3
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
contourpy==1.0.7
cryptography @ file:///tmp/build/80754af9/cryptography_1616769286105/work
csvw==3.1.3
cycler==0.11.0
Cython==0.29.33
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dlinfo==1.2.1
docstring-parser==0.15
einops==0.6.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.10.0
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
h5py==3.8.0
hydra-core==1.3.2
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata==6.1.0
importlib-resources==5.12.0
inflect==6.0.2
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
isodate==0.6.1
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonargparse==4.20.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==8.1.0
jupyter-console==6.6.3
jupyter-core==5.3.0
jupyter-events==0.6.3
jupyter-server==2.5.0
jupyter-server-fileid==0.8.0
jupyter-server-terminals==0.4.4
jupyter-server-ydoc==0.8.0
jupyter-ydoc==0.2.3
jupyterlab==3.6.2
jupyterlab-legos-ui==0.1.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.20.0
jupyterlab-widgets==3.0.6
jupyterthemes==0.20.0
kiwisolver==1.4.4
language-tags==1.2.0
lesscpy==0.15.1
librosa==0.8.0
lightning-utilities==0.8.0
lit==16.0.0
llvmlite==0.39.1
lmdb==1.4.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
natsort==8.3.1
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.3
notebook-shim==0.2.2
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pastel==0.2.1
PeakUtils==1.3.4
pexpect==4.8.0
phonemizer==3.2.1
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==3.1.1
ply==3.11
poethepoet==0.10.0
pooch==1.7.0
praat-parselmouth==0.4.3
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydantic==1.10.6
Pygments==2.14.0
pylatexenc==2.10
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.0.0
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
rdflib==6.3.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.2
rsa==4.9
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
scikit-learn==1.2.2
scipy==1.10.1
segments==2.2.1
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4
sox==1.4.1
srt==3.5.2
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
terminado==0.17.1
tgt==1.4.4
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
torch==2.0.0
torchaudio==2.0.1
torchcde==0.2.5
torchdiffeq==0.2.3
torchdyn @ file:///torchdyn
torchmetrics==0.11.4
torchsde==0.2.5
torchvision==0.15.1
tornado==6.2
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
traitlets==5.9.0
trampoline==0.1.2
triton==2.0.0
typeshed-client==2.2.0
typing-extensions==4.5.0
Unidecode==1.3.6
uri-template==1.2.0
uritemplate==4.1.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.6
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
zipp==3.15.0
